[fix] [test] Fix flaky test ReplicatorTest
